Implement PowerToolsBlock and StronglyTypedBlock
This commit implements the PowerToolsBlock and StronglyTypedBlock
classes, which provide a better way to demarcate and transition between
usages of (1) the PowerTools and Linq-to-XML way of working with Open
XML elements and (2) the strongly typed classes provided by the Open
XML SDK. Those classes leverage extension methods provided by the
PowerToolsBlockExtensions class.

This commit further removes the static ReloadRootElementOnPut property
from the PtOpenXmlExtensions class and its usages from the PutXDocument
methods. It further changes the PutXDocument methods to no longer call
part.RootElement.Reload(). That static property and the invocations of
part.RootElement.Reload() are no longer required in the context of the
PowerToolsBlock, StronglyTypedBlock, and PowerToolsExtensions classes
mentioned above. This commit moves the RemovePowerToolsAnnotations
method to the PowerToolsBlockExtensions class.

Finally, this commit creates or changes the unit tests related to the
PowerToolsBlock, StronglyTypedBlock, and PowerToolsBlockExtensions
classes. It creates The TestsBase class, which contains a helper method
(and more such helper methods in the future) used by unit test classes.